refactor: déplacement logique du cours dans le service associé :recycle:
